+++ Frontend/package.json
@@ -23,5 +23,8 @@
 		"vite": "^5.0.3",
 		"vitest": "^1.2.0"
 	},
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"daisyui": "^4.11.1"
+	}
 }
